Amount needs to use double because of large exponent.

+++ src/main/java/org/joelson/cts/calculator/exploration/AJourneyOfSerenity.java
@@ -17,11 +17,11 @@ private static final GardenState STATE = new GardenState();
 private static final String LEAVES_CURRENCY = "Tea Leaves";
 private static final String CUPS_CURRENCY = "Tea Cups";
 
-private static Amount leaves(float amount) {
+private static Amount leaves(double amount) {
     return new Amount(LEAVES_CURRENCY, amount);
 }
 
-private static Amount cups(float amount) {
+private static Amount cups(double amount) {
     return new Amount(CUPS_CURRENCY, amount);
 }
 
@@ -57,7 +57,7 @@ void main() {
             .addUpgrade("Herbal Medicine", leaves(400_000), 3, false)
             .addUpgradeRequirement("Chagayu")
 
-            .addUpgrade("Ochazuke", leaves(2.5e12f), 100_001, false)
+            .addUpgrade("Ochazuke", leaves(2.5e12), 100_001, false)
             .addUpgradeRequirement("Tea Meals")
             .addUpgradeRequirement("Anti-inflammatory")
             .generator();
@@ -65,200 +65,200 @@ void main() {
     Generator domesticatedTeaPlant = builder.createGenerator("Domesticated Tea Plant", leaves(75_000), 1.15f, leaves(500))
             .addUpgradeRequirement("Cultivation")
 
-            .addUpgrade("Origin Myth", leaves(2e6f), 2, false)
+            .addUpgrade("Origin Myth", leaves(2e6), 2, false)
             .addGeneratorRequirement("Domesticated Tea Plant", 1)
 
-            .addUpgrade("Calm Body and Mind", leaves(1e7f), 3, false)
+            .addUpgrade("Calm Body and Mind", leaves(1e7), 3, false)
             .addUpgradeRequirement("Herbal Medicine")
 
-            .addUpgrade("Digestion", leaves(4e7f), 2.5f, false)
+            .addUpgrade("Digestion", leaves(4e7), 2.5f, false)
             .addUpgradeRequirement("Calm Body and Mind")
 
-            .addUpgrade("Anti-inflammatory", leaves(7e8f), 21, false)
+            .addUpgrade("Anti-inflammatory", leaves(7e8), 21, false)
             .addUpgradeRequirement("Calm Body and Mind")
 
-            .addUpgrade("Weight Management", leaves(2e13f), 1001, false)
+            .addUpgrade("Weight Management", leaves(2e13), 1001, false)
             .addUpgradeRequirement("Cultivation")
             .addUpgradeRequirement("Anti-inflammatory")
             .generator();
 
-    Generator teaPlantation = builder.createGenerator("Tea Plantation", leaves(5e7f), 1.15f, leaves(100_000))
+    Generator teaPlantation = builder.createGenerator("Tea Plantation", leaves(5e7), 1.15f, leaves(100_000))
             .addUpgradeRequirement("Origin Myth")
 
-            .addUpgrade("Pruning", leaves(9e9f), 2.25f, false)
+            .addUpgrade("Pruning", leaves(9e9), 2.25f, false)
             .addGeneratorRequirement("Tea Plantation", 1)
 
-            .addUpgrade("Harvesting", leaves(1.5e11f), 2.5f, false)
+            .addUpgrade("Harvesting", leaves(1.5e11), 2.5f, false)
             .addUpgradeRequirement("Pruning")
 
-            .addUpgrade("Scissors", leaves(4e11f), 2, false)
+            .addUpgrade("Scissors", leaves(4e11), 2, false)
             .addUpgradeRequirement("Harvesting")
 
-            .addUpgrade("Harvesting Machinery", leaves(6e12f), 2, false)
+            .addUpgrade("Harvesting Machinery", leaves(6e12), 2, false)
             .addUpgradeRequirement("Scissors")
 
-            .addUpgrade("Soil Acidity", leaves(1e13f), 5, false)
+            .addUpgrade("Soil Acidity", leaves(1e13), 5, false)
             .addUpgradeRequirement("Harvesting Machinery")
 
-            .addUpgrade("Fertilizer", leaves(9e13f), 2, false)
+            .addUpgrade("Fertilizer", leaves(9e13), 2, false)
             .addUpgradeRequirement("Soil Acidity")
 
-            .addUpgrade("Pest and Disease Control", leaves(9e14f), 4, false)
+            .addUpgrade("Pest and Disease Control", leaves(9e14), 4, false)
             .addUpgradeRequirement("Fertilizer")
 
-            .addUpgrade("Vertical Farming", leaves(6e15f), 3.5f, false)
+            .addUpgrade("Vertical Farming", leaves(6e15), 3.5f, false)
             .addUpgradeRequirement("Pest and Disease Control")
 
-            .addUpgrade("Irrigation System", leaves(1.5e16f), 4, false)
+            .addUpgrade("Irrigation System", leaves(1.5e16), 4, false)
             .addUpgradeRequirement("Harvesting Machinery")
             .addUpgradeRequirement("Vertical Farming")
 
-            .addUpgrade("Mechanical Plucking", leaves(2e17f), 8.5f, false)
+            .addUpgrade("Mechanical Plucking", leaves(2e17), 8.5f, false)
             .addUpgradeRequirement("Irrigation System")
 
-            .addUpgrade("Drone Technology", leaves(8e17f), 11, false)
+            .addUpgrade("Drone Technology", leaves(8e17), 11, false)
             .addUpgradeRequirement("Mechanical Plucking")
 
-            .addUpgrade("Monitoring System", leaves(2e19f), 6, false)
+            .addUpgrade("Monitoring System", leaves(2e19), 6, false)
             .addUpgradeRequirement("Drone Technology") // ? Drone Tech
 
-            .addUpgrade("Storing", leaves(3e8f), 2.5f, false)
+            .addUpgrade("Storing", leaves(3e8), 2.5f, false)
             .addGeneratorRequirement("Tea Plantation", 1)
 
-            .addUpgrade("Drying", leaves(3e10f), 2, false)
+            .addUpgrade("Drying", leaves(3e10), 2, false)
             .addUpgradeRequirement("Storing")
 
-            .addUpgrade("Roasting", leaves(8e11f), 2.25f, false)
+            .addUpgrade("Roasting", leaves(8e11), 2.25f, false)
             .addUpgradeRequirement("Drying")
 
-            .addUpgrade("Fermentation", leaves(3e14f), 3.5f, false)
+            .addUpgrade("Fermentation", leaves(3e14), 3.5f, false)
             .addUpgradeRequirement("Grinding")
             .generator();
 
-    Generator teaEvolution = builder.createGenerator("Tea Evolution", leaves(2e6f), 1.13f, cups(3f))
+    Generator teaEvolution = builder.createGenerator("Tea Evolution", leaves(2e6), 1.13f, cups(3))
             .addUpgradeRequirement("Origin Myth")
 
             .addUpgrade("Tea Contest", cups(1_000), 5, false)
             .addGeneratorRequirement("Tea Evolution", 1)
 
-            .addUpgrade("Silk Road Trade", cups(1e8f), 16, false)
+            .addUpgrade("Silk Road Trade", cups(1e8), 16, false)
             .addGeneratorRequirement("Matcha", 1)
 
-            .addUpgrade("Arabic Shai", cups(2.5e8f), 3.5f, false)
+            .addUpgrade("Arabic Shai", cups(2.5e8), 3.5f, false)
             .addUpgradeRequirement("Silk Road Trade")
 
-            .addUpgrade("Moroccan Atai", cups(4e8f), 3f, false)
+            .addUpgrade("Moroccan Atai", cups(4e8), 3f, false)
             .addUpgradeRequirement("Silk Road Trade")
 
-            .addUpgrade("AI Automation", leaves(9e19f), 20_000_000_000_001f, false)
+            .addUpgrade("AI Automation", leaves(9e19), 20_000_000_000_001f, false)
             .addUpgradeRequirement("Monitoring System")
             .generator();
 
-    Generator matcha = builder.createGenerator("Matcha", cups(1e6f), 1.15f, cups(200))
+    Generator matcha = builder.createGenerator("Matcha", cups(1e6), 1.15f, cups(200))
             .addGeneratorRequirement("Tea Evolution", 1)
 
-            .addUpgrade("Whisking", cups(1.5e7f), 2f, false)
+            .addUpgrade("Whisking", cups(1.5e7), 2f, false)
             .addGeneratorRequirement("Matcha", 1)
 
-            .addUpgrade("Foam Art", cups(5e11f), 10_001, false)
+            .addUpgrade("Foam Art", cups(5e11), 10_001, false)
             .addUpgradeRequirement("Whisking")
             .addUpgradeRequirement("Trade to Europe")
 
-            .addUpgrade("Chanoyu", cups(5e12f), 16, false)
+            .addUpgrade("Chanoyu", cups(5e12), 16, false)
             .addUpgradeRequirement("Foam Art")
 
-            .addUpgrade("Grinding", leaves(4e13f), 2.5f, false)
+            .addUpgrade("Grinding", leaves(4e13), 2.5f, false)
             .addGeneratorRequirement("Matcha", 1)
             .addUpgradeRequirement("Roasting")
             .generator();
 
-    Generator looseLeafTea = builder.createGenerator("Loose-Leaf Tea", cups(3e8f), 1.15f, cups(25_000))
+    Generator looseLeafTea = builder.createGenerator("Loose-Leaf Tea", cups(3e8), 1.15f, cups(25_000))
             .addUpgradeRequirement("Arabic Shai")
             .addUpgradeRequirement("Moroccan Atai")
 
-            .addUpgrade("Steeping", cups(2.5e9f), 3, false)
+            .addUpgrade("Steeping", cups(2.5e9), 3, false)
             .addGeneratorRequirement("Loose-Leaf Tea", 1)
 
-            .addUpgrade("Darye", cups(3e10f), 3.5f, false)
+            .addUpgrade("Darye", cups(3e10), 3.5f, false)
             .addUpgradeRequirement("Steeping")
 
-            .addUpgrade("Trade to Europe", cups(8e10f), 4, false)
+            .addUpgrade("Trade to Europe", cups(8e10), 4, false)
             .addGeneratorRequirement("Loose-Leaf Tea", 1)
 
-            .addUpgrade("Yunnan Pu-erh Tea", leaves(5e15f), 5, false)
+            .addUpgrade("Yunnan Pu-erh Tea", leaves(5e15), 5, false)
             .addGeneratorRequirement("Loose-Leaf Tea", 1)
 
-            .addUpgrade("Tea Brick", leaves(5e16f), 51, false)
+            .addUpgrade("Tea Brick", leaves(5e16), 51, false)
             .addUpgradeRequirement("Yunnan Pu-erh Tea")
             .addUpgradeRequirement("Fermentation")
             .generator();
 
-    Generator infusedTea = builder.createGenerator("Infused Tea", cups(4e13f), 1.15f, cups(2e9f))
+    Generator infusedTea = builder.createGenerator("Infused Tea", cups(4e13), 1.15f, cups(2e9))
             .addUpgradeRequirement("Trade to Europe")
 
-            .addUpgrade("British Tea", cups(1e14f), 9, false)
+            .addUpgrade("British Tea", cups(1e14), 9, false)
             .addGeneratorRequirement("Infused Tea", 1)
 
-            .addUpgrade("Masala Chai", cups(1e15f), 2, false)
+            .addUpgrade("Masala Chai", cups(1e15), 2, false)
             .addGeneratorRequirement("Infused Tea", 1)
 
-            .addUpgrade("Boiling", cups(7.5e15f), 4, false)
+            .addUpgrade("Boiling", cups(7.5e15), 4, false)
             .addUpgradeRequirement("Masala Chai")
 
-            .addUpgrade("High Tea", cups(1.2e17f), 8.5f, false)
+            .addUpgrade("High Tea", cups(1.2e17), 8.5f, false)
             .addUpgradeRequirement("British Tea")
 
-            .addUpgrade("Tea House", cups(8e21f), 2_001, false)
+            .addUpgrade("Tea House", cups(8e21), 2_001, false)
             .addUpgradeRequirement("High Tea")
             .addGeneratorRequirement("Unconventional Tea", 1)
 
-            .addUpgrade("Assam Tea", leaves(4e17f), 2, false)
+            .addUpgrade("Assam Tea", leaves(4e17), 2, false)
             .addUpgradeRequirement("Masala Chai")
 
-            .addUpgrade("Storage Jar", leaves(6e17f), 6, false)
+            .addUpgrade("Storage Jar", leaves(6e17), 6, false)
             .addUpgradeRequirement("Assam Tea")
             .addUpgradeRequirement("Tea Brick")
             .generator();
 
-    Generator unconventionalTea = builder.createGenerator("Unconventional Tea", cups(5e18f), 1.15f, cups(1e15f))
+    Generator unconventionalTea = builder.createGenerator("Unconventional Tea", cups(5e18), 1.15f, cups(1e15))
             .addUpgradeRequirement("Masala Chai")
 
-            .addUpgrade("Iced Tea", cups(4e20f), 2, false)
+            .addUpgrade("Iced Tea", cups(4e20), 2, false)
             .addGeneratorRequirement("Unconventional Tea", 1)
 
-            .addUpgrade("Cold Brew", cups(3e21f), 2.5f, false)
+            .addUpgrade("Cold Brew", cups(3e21), 2.5f, false)
             .addUpgradeRequirement("Iced Tea")
 
-            .addUpgrade("Herbal Tea", cups(7e22f), 6, false)
+            .addUpgrade("Herbal Tea", cups(7e22), 6, false)
             .addUpgradeRequirement("Iced Tea")
 
-            .addUpgrade("Tea Latte", cups(2e23f), 5, false)
+            .addUpgrade("Tea Latte", cups(2e23), 5, false)
             .addUpgradeRequirement("Herbal Tea")
 
-            .addUpgrade("Tea Cocktail", cups(1.5e24f), 201f, false)
+            .addUpgrade("Tea Cocktail", cups(1.5e24), 201f, false)
             .addUpgradeRequirement("Herbal Tea")
 
-            .addUpgrade("Bubble Tea", cups(3e25f), 6, false)
+            .addUpgrade("Bubble Tea", cups(3e25), 6, false)
             .addUpgradeRequirement("Tea Latte")
 
-            .addUpgrade("Tea Bag", leaves(6e18f), 2.5f, false)
+            .addUpgrade("Tea Bag", leaves(6e18), 2.5f, false)
             .addGeneratorRequirement("Unconventional Tea", 1)
             .addUpgradeRequirement("Storage Jar")
 
-            .addUpgrade("Vacuum Sealer", leaves(3e19f), 2.5f, false)
+            .addUpgrade("Vacuum Sealer", leaves(3e19), 2.5f, false)
             .addUpgradeRequirement("Tea Bag")
             .generator();
 
-    Generator virtualTea = builder.createGenerator("Virtual Tea", cups(5e22f), 1.15f, cups(4e18f))
+    Generator virtualTea = builder.createGenerator("Virtual Tea", cups(5e22), 1.15f, cups(4e18))
             .addUpgradeRequirement("AI Automation")
 
-            .addUpgrade("Tea Simulator", cups(5e23f), 4f, false)
+            .addUpgrade("Tea Simulator", cups(5e23), 4f, false)
             .addGeneratorRequirement("Virtual Tea", 1)
 
-            .addUpgrade("Online Tea Cermony", cups(1e25f), 6f, false)
+            .addUpgrade("Online Tea Cermony", cups(1e25), 6f, false)
             .addGeneratorRequirement("Virtual Tea", 1)
 
-            .addUpgrade("Shared Serenity", cups(1e26f), 2, false)
+            .addUpgrade("Shared Serenity", cups(1e26), 2, false)
             .addUpgradeRequirement("Tea Simulator")
             .addUpgradeRequirement("Online Tea Cermony")
             .generator();
